Add files via upload
Login APP using the MVC model and the JSTL 

this is the Dynamic web project which gets the username and password inputs from the 
login.jsp and sends it for the validation
after validating it create the new response and send to the browser but we are using the request
Dispatcher, it will get the request on the server site itself and send the response to the user without changing the actual URL. we are sending the user object instead of sending the data using the session or the application objects
and from that user object we fetch the data and display to produce the dynamic result. we are using the JSTL because 

in normal JSP file lot of java code is written by the coder it increases the complexity . to avoid that we use JSTL 

JSTL is easy to use. ( we use JSTL because the web developers and the web designer are two different persons.